+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- fj-doc-freemarker, html handler now renders fore color for phrase.
+
 ## [8.17.7] - 2025-11-20
 
 ### Added