+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.17.8] - 2025-11-26
+
 ### Added
 
 - fj-doc-base, new attribute 'class' for element <para>, <phrase>, <h> (same as html class, onlt for freemarker html)